+++ src/prime_rl/trainer/sft/train.py
@@ -240,7 +240,9 @@ def train(config: SFTTrainerConfig):
                 # Forward pass
             logger.debug("Starting forward pass")
             with maybe_record_function("forward"), maybe_activation_offloading(config.model.ac_offloading):
-                logits = forward(model, input_ids, position_ids)
+                out = forward(model, input_ids, position_ids)
+
+            logits = out.logits
             B, L, V = logits.shape
 
             # Compute loss