+++ tests/unit/train/test_model.py
@@ -109,6 +109,8 @@ def test_moe_custom_impl():
     config = ModelConfig(name="PrimeIntellect/GLM-0.5B", attn="sdpa", impl="custom", moe_use_grouped_mm=False)
     model = get_model(config)
     model = model.to("cuda")
+    # we need to wrap the lm head as custom forward only works with it, this is done in setup_model
+    model.wrap_lm_head(chunk_size=None)
     with torch.autocast("cuda", dtype=torch.bfloat16):
         inputs_ids = torch.randint(0, 100, (BS, SEQ_LEN)).to("cuda")
         outputs = model(input_ids=inputs_ids).logits
@@ -121,6 +123,8 @@ def test_moe_custom_impl():
 def test_model_forward_custom_impl(model_name):
     config = ModelConfig(name=model_name, impl="custom", attn="sdpa")
     model = get_model(config)
+    # we need to wrap the lm head as custom forward only works with it, this is done in setup_model
+    model.wrap_lm_head(chunk_size=None)
     model = model.to("cuda")
     with torch.autocast("cuda", dtype=torch.bfloat16):
         inputs_ids = torch.randint(0, 100, (BS, SEQ_LEN)).to("cuda")